pyproject.toml
[project]
name = "test_auth_app"
version = "0.1.0"
description = ""
authors = [
    {name="Evgeniy Anistratov", email="evgeneanistratov@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
django = "^5.2.8"
pyjwt = "^2.10.1"
djangorestframework = "^3.16.1"
envparse = "^0.2.0"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.11"
bcrypt = "^5.0.0"
